Micro-optimization finished
This micro-optimization strategy was implemented for the Publish View. This view had a lot of TextEditingControllers, which were not disposed when the widget was removed from the widget tree. This caused an innecesary allocation of them in memory, which affected the performance as the space was smaller. By adding the dispose() function, all the instances of those controllers were removed from memory as soon as the widget was used any more. Improving the performance and reducing the amount of garbage collections needed