Chat Implementation finished
A new feature was added so that the users were able to talk between them. They can create a chat group or to search for an existing one. In these groups they are supposed to talk about some products they are interested in, and also offer the ones they are selling. It was implemented the eventual connectivity protection for this view, which is to store the messages sent while not being connected in the cache, so that when the connection was restored, the messages where sent automatically.